+++ projects/Mallard/src/components/layout/ui/sizing/with-breakpoints.tsx
@@ -1,25 +1,7 @@
-import React, { useState, ReactNode, FunctionComponent, useEffect } from 'react'
-import { Dimensions, ScaledSize } from 'react-native'
+import React, { ReactNode, FunctionComponent } from 'react'
+import { ScaledSize } from 'react-native'
 import { getClosestBreakpoint, BreakpointList } from 'src/theme/breakpoints'
-
-const useDimensions = (): ScaledSize => {
-    const [dimensions, setDimensions] = useState(Dimensions.get('window'))
-    useEffect(() => {
-        const listener = (
-            ev: Parameters<
-                Parameters<typeof Dimensions.addEventListener>[1]
-            >[0],
-        ) => {
-            setDimensions(ev.screen)
-        }
-        Dimensions.addEventListener('change', listener)
-        return () => {
-            Dimensions.removeEventListener('change', listener)
-        }
-    }, [])
-
-    return dimensions
-}
+import { useDimensions } from 'src/hooks/use-screen'
 
 const WithBreakpoints: FunctionComponent<{
     children: BreakpointList<(d: ScaledSize) => ReactNode>
