+++ projects/Mallard/src/components/login/login-layout.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { StyleSheet, View, KeyboardAvoidingView } from 'react-native'
 import { metrics } from 'src/theme/spacing'
 import { color } from 'src/theme/color'
-import { useInsets } from 'src/hooks/use-insets'
+import { useInsets } from 'src/hooks/use-screen'
 import { Button } from '../button/button'
 import { TitlepieceText, UiBodyCopy } from '../styled-text'
 import { Spinner } from '../spinner'
